Still a work in progress but it is getting there
This version of the log investigator will read the log file and get the sessions, average session length, commands issued and time between commands (time between commands is still a bit buggy) there are also totals at the bottom of the csv file. And now the log investigator outputs to a csv file that is called Output.csv. So I would recommend opening that in excel. Report any bugs that you find in the issues section in this repo